+++ packages/pluggableWidgets/datagrid-date-filter-web/src/DatagridDateFilter.editorConfig.ts
@@ -4,6 +4,7 @@ import {
     ContainerProps,
     datePickerIcon,
     datePickerIconDark,
+    emptyIcon,
     equalsIcon,
     equalsIconDark,
     greaterThanEqualIcon,
@@ -13,6 +14,7 @@ import {
     hidePropertiesIn,
     hidePropertyIn,
     ImageProps,
+    notEmptyIcon,
     notEqualIcon,
     notEqualIconDark,
     Properties,
@@ -146,5 +148,9 @@ function getSvgContent(type: DefaultFilterEnum, isDarkMode: boolean): string {
             return isDarkMode ? smallerThanIconDark : smallerThanIcon;
         case "smallerEqual":
             return isDarkMode ? smallerThanEqualIconDark : smallerThanEqualIcon;
+        case "empty":
+            return emptyIcon;
+        case "notEmpty":
+            return notEmptyIcon;
     }
 }