+++ packages/pluggableWidgets/datagrid-number-filter-web/typings/DatagridNumberFilterProps.d.ts
@@ -7,7 +7,7 @@ import { CSSProperties } from "react";
 import { ActionValue, DynamicValue, EditableValue } from "mendix";
 import { Big } from "big.js";
 
-export type DefaultFilterEnum = "greater" | "greaterEqual" | "equal" | "notEqual" | "smaller" | "smallerEqual";
+export type DefaultFilterEnum = "greater" | "greaterEqual" | "equal" | "notEqual" | "smaller" | "smallerEqual" | "empty" | "notEmpty";
 
 export interface DatagridNumberFilterContainerProps {
     name: string;