+++ packages/pluggableWidgets/datagrid-date-filter-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   We added the options to filter for empty and non-empty values.
+
 ## [2.3.1] - 2022-04-13
 
 ### Fixed